SCRIPTS: fix Temenos avatars (they are all named "Mystic_Avatar" in core)
